+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Fix page numbering in `BakeIndex`
 * Change EOC titles in `world-history`
 * Modify `BakeNumberedNotes.v1` to bake notes within Appendix
 * Add `appendix` to `contemporary-math`